found an interesting bug (fixed)
when making a list variable = to another list variable does not simply copy the contents of the original list.  It creates a reference essentially passing the variable itself.  When you update the reference variable, it also updates the original variable.  This was happening with my deck of cards.  Solution was to change "deck=standarddeck" to "deck=standarddeck[:]"  This version has that fix.